Some fixes + IFGUARDS

+++ src/Common/Coordinate.h
@@ -10,9 +10,9 @@
 #define COORDINATE_H
 class Coordinate {
 	public:
-		Coordinate( double x, double y, double z )	:  x(x) , y(y) , z(z) {}
-	double x;	// per adesso senza getter e setter
-	double y;
-	double z;	
+		Coordinate( float x, float y, float z )	:  x(x) , y(y) , z(z) {}
+	float x;	// per adesso senza getter e setter
+	float y;
+	float z;	
 };
 #endif
